core-service/requirements.txt
# requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
